requirements.txt
streamlit
tensorflow-cpu==2.12.0
numpy
pillow
h5py